+++ test/integration/smoke/test_projects.py
